Lecture07. Refactor. Create separate classes for Connection, and Selects
