+++ src/main/java/eu/senla/dto/dbResponse/AdminRequestDataFromDB.java
@@ -1,0 +1,7 @@
+package eu.senla.dto.dbResponse;
+
+public record AdminRequestDataFromDB(
+        int staffId,
+        String lastName,
+        String passport) {
+}
