+++ src/main/java/eu/senla/dto/dbResponse/UserRequestDataFromDB.java
@@ -1,0 +1,8 @@
+package eu.senla.dto.dbResponse;
+
+public record UserRequestDataFromDB(
+        int applicationId,
+        int applicantId,
+        String lastName,
+        String firstName) {
+}
